+++ packages/jest-reporters/src/index.ts
@@ -6,6 +6,8 @@
  */
 
 import getResultHeader from './getResultHeader';
+import getSnapshotStatus from './getSnapshotStatus';
+import getSnapshotSummary from './getSnapshotSummary';
 import {
   formatTestPath,
   getSummary,
@@ -14,12 +16,15 @@ import {
   trimAndFormatPath,
 } from './utils';
 
-export type {Config} from '@jest/types';
 export type {
   AggregatedResult,
   SnapshotSummary,
+  Test,
+  TestCaseResult,
+  TestContext,
   TestResult,
 } from '@jest/test-result';
+export type {Config} from '@jest/types';
 export {default as BaseReporter} from './BaseReporter';
 export {default as CoverageReporter} from './CoverageReporter';
 export {default as DefaultReporter} from './DefaultReporter';
@@ -28,15 +33,16 @@ export {default as SummaryReporter} from './SummaryReporter';
 export {default as VerboseReporter} from './VerboseReporter';
 export {default as GitHubActionsReporter} from './GitHubActionsReporter';
 export type {
-  Context,
   Reporter,
   ReporterOnStartOptions,
+  ReporterContext,
   SummaryOptions,
-  Test,
 } from './types';
 export const utils = {
   formatTestPath,
   getResultHeader,
+  getSnapshotStatus,
+  getSnapshotSummary,
   getSummary,
   printDisplayName,
   relativePath,